test/requirements.txt
# this is the requirements dedicated to tests.
-r ../requirements.txt
unittest2
mock
coveralls
nose-cov
coverage
nose
pylint
pep8
pep257
requests_mock